+++ cli/types.ts
@@ -37,6 +37,10 @@ export type CanisterConfig = {
   args?: string[];
   candid?: string;
   initArg?: string;
+  "check-stable"?: {
+    path: string;
+    skipIfMissing?: boolean;
+  };
 };
 
 export type Dependencies = Record<string, Dependency>;